requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
sse-starlette
requests
pydantic
pdf2image
Pillow
python-dotenv
